==> Julia/Notebooks/3. DataFrames/3. Working with text files.ipynb <==
# Working with text files


In this notebook we will show how one can interact with CSV files when working with DataFrames.

using DataFrames

using CSV

using Arrow

using Statistics

First we download the data set we will work with and save it as auto.txt file in a current working directory.

download("https://archive.ics.uci.edu/ml/machine-learning-databases/auto-mpg/auto-mpg.data-original",
         "auto.txt")

Let us check how the file looks inside using the `readlines` function:

readlines("auto.txt")

For this exercise we have chosen a typical file, which in practice means that things are not trivial.
* first we note that it has no header with column names.
* second, we see that the last column is tab separated, while earlier columns are separated with varying number of spaces
* finally we see that missing values are encoded by "NA" in this file

We will show several options how the file can be parsed to a `DataFrame`.
The first one is to replace tabs with spaces in the source file, and then load it using `CSV.File` command.

We start by getting the contents of the file into a single string.

raw_str = read("auto.txt", String)

Now we replace all tabs in this string by spaces

(note that in general it is not a safe operation as theoretically if you had columns that are strings they might have contained quoted tabs; fortunately in this case they do not have them so we are safe).

str_no_tab = replace(raw_str, '\t'=>' ')

Finally we create an `IOBuffer` backed by the string we have just created.

io = IOBuffer(str_no_tab)

You can think of variable `io` as an in-memory I/O stream. Therefore we can pass this stream to the `CSV.File` function to read it as-if it were a CSV file. Note that in the options we choose that:
* the delimiter is space
* we ignore repeated (consecutive) occurences of the delimiter (so we correctly handle our file which has columns padded by spaces)
* we explicitly pass column names via `header` keyword argument
* we specify that missing values are represented using `"NA"` string in our file

Finally note that we pass the result of `CSV.File` operation to a `DataFrame` constructor using `|>`.

df1 = CSV.File(io,
               delim=' ',
               ignorerepeated=true,
               header=[:mpg, :cylinders, :displacement, :horsepower,
                       :weight, :acceleration, :year, :origin, :name],
               missingstring="NA") |>
      DataFrame

Note that not all columns of the data frame have been displayed and we see 30 rows (they do not fit on my screen).

It is easy to change the default maximum width and height of the output by setting appropriate values in `ENV` dictionary.

ENV["COLUMNS"], ENV["LINES"] = 200, 15

df1

Now the display is much nicer.

Let us discuss an alternative way to read in the original file.
This time we will first read the data in directly from the file.

df_raw = CSV.File("auto.txt", header=[:metrics, :name]) |> DataFrame

Note that this time CSV.jl auto-detected that tab is the right delimiter to split the columns

(it was the only delimiter that produced consistent number of columns).

We will split `:metrics` column manually now

str_metrics = split.(df_raw.metrics)

Now let us create an empty `df1_2` data frame that we will populate with appropriate columns.
The pattern we use here is typical when you e.g. perform repeated computations whose results you want to store in a `DataFrame`.

df1_2 = DataFrame([col => Float64[] for
                  col in [:mpg, :cylinders, :displacement, :horsepower, :weight, :acceleration, :year, :origin]])

Now we have a data frame that has 8 columns and 0 rows. It accepts floating point values. However in columns `:mpg` and `:horsepower` we have to allow the data frame to hold missing values. We do it using `allowmissing!` function

allowmissing!(df1_2, [:mpg, :horsepower])

Note that the element type of columns `:mpg` and `:horsepower` changed to `Float64?` which signals that that columns allows missing values.

Now we are ready to populate our data frame.

for row in str_metrics
    push!(df1_2, [v == "NA" ? missing : parse(Float64, v) for v in row])
end

df1_2

finally, let us add a column `:name` from the `df_raw` data frame

df1_2.name = df_raw.name

df1_2

Before we move forward we should stress one very important thing that is related to `df1_2.name = df_raw.name` assignment. After this operation columns `:name` in both `df1_2` and `df_raw` are the same objects.
We can easily check it:

df1_2.name === df_raw.name

Such behavior is allowed for performance reasons. In this case we accepted it as we want to discard `df_raw` data frame and not use it in later analysis.
However, in general it would be safer to create a copy of `:name` column when assigning it ti `df1_2` data frame.

It can be achieved like this:
```
df1_2[:, :name] = df_raw.name
```
or like this
```
df1_2.name = df_raw[:, :name]
```
We could also write:
```
df1_2[:, :name] = df_raw[:, :name]
```
but this time there would be one unnecessary copy made (one when reading the data from `df_raw` the other when writing data to `df1_2`)

We can check that `df1` and `df1_2` data frames are equal using `isequal` funtion, so that we ended up with identical data frames.

isequal(df1_2, df1)

Note that as the data frames contain missing values comparing them with `missing` would produce `missing`.

df1_2 == df1

Let us investigate yet another way to create our data frame, this time in one-shot (this line is more difficult to understand, but it shows you the power that DataFrames.jl gives you when working with data, in part 4 of the tutorial we will give some more examples of supported transformations):

df1_3 = select(df_raw,
               :metrics =>
               ByRow(x -> something.(tryparse.(Float64, split(x)), missing)) =>
               [:mpg, :cylinders, :displacement, :horsepower, :weight, :acceleration, :year, :origin],
               :name)

Let us understand what is going on in the above expression.

The general rule os specifying transformations in `select` (also `transform` and `combine`) is:

`source_columns => transformation => destination_columns`

In this case we take `:metrics` column from the source data frame. We process it element by element,
which is signaled by `ByRow` (you can think of this as broadcasting).
Now each element is first `split` and then we try parsing it as `Float64`.
If we fail `tryparse` returns `nothing`, which we convert to `missing` using the `something` function.
As a result we obtain a vector of `Float64` or `Missing` values.
Since the last element of our transformation is a vector of culumn names this vector is expanded into
that many columns.

Finally we add the `:name` column from the source data frame

As above we check that the result is the same as `df1`.

isequal(df1_3, df1)

We can easily count the number of missing values in the `df1` data frame using the `eachcol` function that returns the iterator over columns of the data frame:

sum(count(ismissing, col) for col in eachcol(df1))

we could alternatively transform our data frame to a `Matrix` and use `count` on it (it would be slower though as unnecessary copies of data would be performed):

count(ismissing, Matrix(df1))

or if you like iterators:

count(ismissing, Iterators.flatten(eachcol(df1)))

Alternatively could use the `mapcols` function to get the number of missings per column:

mapcols(x -> count(ismissing, x), df1)

also it is easy to find the rows containing missing values using the `filter` function:

filter(row -> any(ismissing, row), df1)

Assume we are interested in the brand of each car. We can extract it from `:name` column using broadcasting like this:

df1.brand = first.(split.(df1.name))

df1

Earlier we have shown how one can manually find rows that have missing values. A common operation is the reverse - i.e. selecting only rows that do not contain missing values. This can be achieved like this:

df2 = dropmissing(df1)

Now let us find all rows that correspond to `"saab"` brand. You can do it in two ways, either indexing or using `filter` function:

df2[df2.brand .== "saab", :]

filter(:brand => ==("saab"), df2)

Note that the `:brand => ==("saab")` syntax means that we take elements of `:brand` column and pass it to `==("saab")` function.

Now `==("saab")` is just a shorthand for `x -> x == "saab"`.

Alternatively we could do the filtering operation in the following way (this is a bit slower but you might find it more readable):

filter(row -> row.brand == "saab", df2)

To finish this part of the tutorial let us save the `df2` file to auto2.csv and auto2.arrow files. We will use them later in the next parts of the course.

CSV.write("auto2.csv", df2)

Arrow.write("auto2.arrow", df2)

Let us just quickly inspect what we have written to disk in auto2.csv (auto2.arrow file is binary) before we finish:

readlines("auto2.csv")

Note that CSV.jl by default used a comma to separate the fields in our file and written a header in the first row of the file.
