==> Julia/Notebooks/3. DataFrames/4. Grouping data frames.ipynb <==
# Working with groups of rows of a data frame



using DataFrames

using CSV

# Pkg.add("Arrow")
using Arrow

using Statistics

using FreqTables

using Pipe

This part of the tutorial assumes that you have run *3. Working with text files* part so that you have auto2.csv and auto2.arrow files in your working directory.

ENV["LINES"], ENV["COLUMNS"] = 15, 200

df = CSV.File("auto2.csv") |> DataFrame

df2 = Arrow.Table("auto2.arrow") |> DataFrame

df == df2

In general using Arrow.jl is a preferred way to store your data frames, but note that in order to ensure maximum speed it uses its own `AbstractVector` type:

df2.mpg

To materialize columns into standard `Vector`s just `copy` a data frame:

df3 = copy(df2)

df3.mpg

You can find more discussion about using Arrow.jl with DataFrames.jl [here](https://bkamins.github.io/julialang/2020/11/06/arrow.html).

We want to group our data frame by `:brand` column:

gdf = groupby(df, :brand)

Note that it is easy (and fast) to find an appropriate group by passing a `Tuple` of values of grouping columns (in our case it is just one column)

gdf[("ford",)]

Using the `combine` function we can easily calculate some aggregates by group:

brand_mpg = combine(gdf, :mpg => mean)

The `:mpg => mean` syntax means that we pass a column `:mpg` to be processed by the `mean` function for each group.

We could have added a target column name to the generated column in the following way:

combine(gdf, :mpg => mean => :mean_mpg)

Now let us see how one can sort a data frame. In order to see alll rows of a data frame we temporarily change the number of rows printed.

ENV["LINES"] = 50

sort!(brand_mpg, :mpg_mean, rev=true) # sort! performs an in-place operation

A typical data cleaning task is to check its consistency. In this case it would be making sure that each brand has a unique origin. We will do it in several ways.

First we use FreqTables.jl:

freqtable(df, :brand, :origin)

Now let us create a chain of operations using Pipe.jl. Here `_` denotes the position of the argument passed from the previous step in the chain.

orig_brand = @pipe df |>
                   groupby(_, :brand) |>
                   combine(_, :origin => x -> length(unique(x)))

Let us make sure that we always get exactly one origin:

extrema(orig_brand.origin_function)

We also could have grouped our data frame by two columns:

origin_brand2 = @pipe df |>
                      groupby(_, [:origin, :brand]) |>
                      combine(_, nrow)

Note that passing `nrow` as a combining operation gives us the number of rows per group.

Now we can do the `unstack` operation on the result to get a similar table that we got with `freqtable`:

origin_vs_brand = unstack(origin_brand2, :brand, :origin, :nrow)

The only difference is that missing combinations are given as `missing` and `freqtable` produced `0` there. It is easy to fix using broadcasting:

coalesce.(origin_vs_brand, 0)

Note that column names of a data frame can be any sequence of characters as in this case:

names(origin_vs_brand)

propertynames(origin_vs_brand)

In such situations it is easiest to refer to such columns using strings, as `Symbol`s are not super easy to write when they have to represent sequences that are not allowed as variable names as you can see above.

origin_vs_brand."1.0"

Finally let us show that columns of a data frame can hold any data type.

origin_brand3 = @pipe df |>
                      groupby(_, :origin) |>
                      combine(_, :brand => x -> Ref(unique(x)))

Note that we used `Ref` to protect the return value of our function to be broadcasted (this is the same trick as in standard broadcasting in Julia). Otherwise we would get:

@pipe df |> groupby(_, :origin) |> combine(_, :brand => unique)

Now we check if the groups are pairwise disjoint (this is a bit tricky, but it is much fun to be able to write such code in Julia):

for i in axes(origin_brand3, 1)
    for j in i+1:nrow(origin_brand3)
        println("$i vs $j => ", intersect(origin_brand3.brand_function[[i,j]]...))
    end
end

We can also easily flatten the `origin_brand3` data frame so that `brand_function` column is expanded to span multiple rows (the case we have seen above when we omitted writing `Ref` in `combine`)

flatten(origin_brand3, :brand_function)
